mainFolder/test2_faceD/requirements.txt
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
joblib==1.3.2
pillow==10.2.0
face-recognition==1.3.0
pyttsx3==2.90
matplotlib==3.8.2
seaborn==0.13.2
ultralytics==8.1.2
python-dotenv==1.0.1
plotly==5.18.0
playsound==1.3.0 